refactor:portal 모델 최적화
